🎨 共通コンポーネントの切り出し
